refactor code to use ProductManager instead of profuct repository
